chore: inline text message consumer
